+++ src/settings.ts
@@ -7,6 +7,7 @@ import DataViewObjectsParser = dataViewObjectsParser.DataViewObjectsParser;
 export interface IVisualSettings {
     view: ViewSettings;
     mermaid: MermaidSettings;
+    mermaidDebug: MermaidDebugSettings;
     font: FontSettings;
     markdown: MarkdownSettings;
 }
@@ -14,6 +15,7 @@ export interface IVisualSettings {
 export class VisualSettings extends DataViewObjectsParser implements IVisualSettings {
     public view: ViewSettings = new ViewSettings();
     public mermaid: MermaidSettings = new MermaidSettings();
+    public mermaidDebug: MermaidDebugSettings = new MermaidDebugSettings();
     public font: FontSettings = new FontSettings();
     public markdown: MarkdownSettings = new MarkdownSettings();
 }
@@ -24,19 +26,23 @@ export class ViewSettings {
 }
 
 export class MermaidSettings {
+    public flowchartOrientation: string = "default";
+    public maxEdges: number = 30000;
+    public securityLevel: string = "loose";
+}
+
+export class MermaidDebugSettings {
     public htmlLabels: boolean = true;
     public markdownAutoWrap: boolean = true;
-    public securityLevel: string = "loose";
-    public maxEdges: number = 30000;
     public convertBrToNewline: boolean = true;
     public autoBacktickLabels: boolean = true;
     public preserveLineBreaksCSS: boolean = true;
 }
 
 export class FontSettings {
-    public fontFamily: string = "Segoe UI";
-    public headingFontSize: number = 20;
-    public bodyFontSize: number = 11;
+    public fontFamily: string = "DIN";
+    public headingFontSize: number = 14;
+    public bodyFontSize: number = 9;
     public mermaidFontSize: number = 14;
 }